refactor: replace with @varlet/shared
