+++ packages/varlet-ui/src/back-top/BackTop.vue
@@ -23,7 +23,8 @@ import { defineComponent, ref, onMounted, onBeforeUnmount } from 'vue'
 import VarButton from '../button'
 import VarIcon from '../icon'
 import { props } from './props'
-import { isString, easeInOutCubic, throttle, isObject } from '../utils/shared'
+import { isString, throttle, isObject } from '@varlet/shared'
+import { easeInOutCubic } from '../utils/shared'
 import { getScrollTop, getScrollLeft, scrollTo, getParentScroller, toPxNum, toSizeUnit } from '../utils/elements'
 import { call, createNamespace } from '../utils/components'
 import type { Ref, TeleportProps } from 'vue'
