+++ packages/varlet-ui/src/bottom-navigation/BottomNavigation.vue
@@ -25,7 +25,7 @@ import { defineComponent, ref, computed, onMounted, onUpdated, watch } from 'vue
 import { props } from './props'
 import { useBottomNavigationItems } from './provide'
 import { createNamespace, call } from '../utils/components'
-import { isNumber } from '../utils/shared'
+import { isNumber } from '@varlet/shared'
 import type { BottomNavigationProvider } from './provide'
 import type { BottomNavigationItemProvider } from '../bottom-navigation-item/provide'
 import type { Ref, ComputedRef } from 'vue'
