+++ packages/varlet-ui/src/checkbox-group/CheckboxGroup.vue
@@ -14,7 +14,7 @@ import { props } from './props'
 import { useValidation, createNamespace, call } from '../utils/components'
 import { useCheckboxes } from './provide'
 import { useForm } from '../form/provide'
-import { uniq } from '../utils/shared'
+import { uniq } from '@varlet/shared'
 import type { ComputedRef } from 'vue'
 import type { ValidateTriggers } from './props'
 import type { CheckboxGroupProvider } from './provide'
