+++ packages/varlet-ui/src/collapse/Collapse.vue
@@ -9,7 +9,7 @@ import { computed, defineComponent, nextTick, watch } from 'vue'
 import { useCollapseItem } from './provide'
 import { props } from './props'
 import { call, createNamespace } from '../utils/components'
-import { isArray } from '../utils/shared'
+import { isArray } from '@varlet/shared'
 import type { ComputedRef } from 'vue'
 import type { CollapseItemProvider } from '../collapse-item/provide'
 import type { CollapseProvider } from './provide'
