+++ packages/varlet-ui/src/date-picker/DatePicker.vue
@@ -79,7 +79,7 @@ import MonthPickerPanel from './src/month-picker-panel.vue'
 import YearPickerPanel from './src/year-picker-panel.vue'
 import DayPickerPanel from './src/day-picker-panel.vue'
 import { props, MONTH_LIST, WEEK_HEADER } from './props'
-import { isArray, toNumber } from '../utils/shared'
+import { isArray, toNumber } from '@varlet/shared'
 import { nextTickFrame } from '../utils/elements'
 import { createNamespace, call } from '../utils/components'
 import { pack } from '../locale'
