+++ packages/varlet-ui/src/date-picker/src/day-picker-panel.vue
@@ -44,7 +44,7 @@ import PanelHeader from './panel-header.vue'
 import VarButton from '../../button'
 import { defineComponent, ref, computed, watch, onMounted, reactive } from 'vue'
 import { WEEK_HEADER } from '../props'
-import { toNumber } from '../../utils/shared'
+import { toNumber } from '@varlet/shared'
 import { createNamespace } from '../../utils/components'
 import { pack } from '../../locale'
 import type { Ref, ComputedRef, UnwrapRef, PropType, RendererNode } from 'vue'
