+++ packages/varlet-ui/src/icon/Icon.vue
@@ -23,7 +23,7 @@
 
 <script lang="ts">
 import { defineComponent, watch, ref, nextTick } from 'vue'
-import { isURL, toNumber } from '../utils/shared'
+import { isURL, toNumber } from '@varlet/shared'
 import { props } from './props'
 import { toSizeUnit } from '../utils/elements'
 import type { Ref } from 'vue'
