+++ packages/varlet-ui/src/input/Input.vue
@@ -128,7 +128,7 @@ import VarFormDetails from '../form-details'
 import VarIcon from '../icon'
 import { defineComponent, getCurrentInstance, ref, computed, nextTick, onMounted } from 'vue'
 import { props } from './props'
-import { isEmpty } from '../utils/shared'
+import { isEmpty } from '@varlet/shared'
 import { useValidation, createNamespace, call } from '../utils/components'
 import { useForm } from '../form/provide'
 import type { Ref, ComputedRef } from 'vue'
