+++ packages/varlet-ui/src/lazy/index.ts
@@ -1,5 +1,6 @@
 import { getAllParentScroller, inViewport } from '../utils/elements'
-import { createCache, removeItem, throttle } from '../utils/shared'
+import { removeItem, throttle } from '@varlet/shared'
+import { createCache } from '../utils/shared'
 import type { App, Directive, Plugin, DirectiveBinding } from 'vue'
 import type { CacheInstance } from '../utils/shared'
 import { call } from '../utils/components'
