+++ packages/varlet-ui/src/list/List.vue
@@ -29,7 +29,8 @@ import Ripple from '../ripple'
 import { defineComponent, onMounted, onUnmounted, ref, nextTick } from 'vue'
 import { getParentScroller, toPxNum } from '../utils/elements'
 import { props } from './props'
-import { isNumber, dt } from '../utils/shared'
+import { isNumber } from '@varlet/shared'
+import { dt } from '../utils/shared'
 import { createNamespace, call } from '../utils/components'
 import { pack } from '../locale'
 import type { Ref } from 'vue'
