+++ packages/varlet-ui/src/pagination/Pagination.vue
@@ -108,7 +108,7 @@ import VarCell from '../cell'
 import VarInput from '../input'
 import { defineComponent, ref, computed, watch } from 'vue'
 import { props } from './porps'
-import { isNumber, toNumber } from '../utils/shared'
+import { isNumber, toNumber } from '@varlet/shared'
 import { pack } from '../locale'
 import type { ComputedRef, Ref } from 'vue'
 import type { Range } from './porps'
