+++ packages/varlet-ui/src/picker/Picker.vue
@@ -91,7 +91,8 @@ import VarButton from '../button'
 import VarPopup from '../popup'
 import { defineComponent, watch, ref, computed, Transition, toRaw } from 'vue'
 import { props } from './props'
-import { isArray, dt } from '../utils/shared'
+import { isArray } from '@varlet/shared'
+import { dt } from '../utils/shared'
 import { toPxNum, getTranslate } from '../utils/elements'
 import { pack } from '../locale'
 import type { Ref, ComputedRef, ComponentPublicInstance } from 'vue'
