+++ packages/varlet-ui/src/pull-refresh/PullRefresh.vue
@@ -24,7 +24,7 @@ import VarIcon from '../icon'
 import { defineComponent, ref, computed, watch, onMounted } from 'vue'
 import { getParentScroller, getScrollTop } from '../utils/elements'
 import { props } from './props'
-import { toNumber } from '../utils/shared'
+import { toNumber } from '@varlet/shared'
 import { createNamespace } from '../utils/components'
 import type { Ref } from 'vue'
 import type { RefreshStatus } from './props'
