+++ packages/varlet-ui/src/select/Select.vue
@@ -118,7 +118,7 @@ import VarMenu from '../menu'
 import VarChip from '../chip'
 import VarFormDetails from '../form-details'
 import { computed, defineComponent, ref, watch, nextTick } from 'vue'
-import { isArray, isEmpty } from '../utils/shared'
+import { isArray, isEmpty } from '@varlet/shared'
 import { props } from './props'
 import { useValidation, createNamespace, call } from '../utils/components'
 import { useOptions } from './provide'
