+++ packages/varlet-ui/src/space/Space.tsx
@@ -2,7 +2,7 @@ import { defineComponent, VNodeChild, Fragment, VNode, Comment } from 'vue'
 import { internalSizeValidator, props } from './props'
 import type { SpaceInternalSize, SpaceSize } from './props'
 import { toPxNum } from '../utils/elements'
-import { isArray } from '../utils/shared'
+import { isArray } from '@varlet/shared'
 import '../styles/common.less'
 import './space.less'
 import { call, createNamespace } from '../utils/components'
