+++ packages/varlet-ui/src/tabs/Tabs.vue
@@ -46,7 +46,8 @@ import VarSticky from '../sticky'
 import { defineComponent, watch, ref, computed, Transition, onMounted, onUnmounted } from 'vue'
 import { props } from './props'
 import { useTabList } from './provide'
-import { isNumber, linear } from '../utils/shared'
+import { isNumber } from '@varlet/shared'
+import { linear } from '../utils/shared'
 import { toSizeUnit, scrollTo, doubleRaf } from '../utils/elements'
 import type { Ref, ComputedRef } from 'vue'
 import type { TabsProvider } from './provide'
