+++ packages/varlet-ui/src/time-picker/TimePicker.vue
@@ -58,7 +58,7 @@ import { computed, defineComponent, reactive, ref, watch } from 'vue'
 import dayjs from 'dayjs/esm'
 import Clock from './clock.vue'
 import { props, hoursAmpm, hours24 } from './props'
-import { toNumber } from '../utils/shared'
+import { toNumber } from '@varlet/shared'
 import { createNamespace, call } from '../utils/components'
 import { getNumberTime, getIsDisableMinute, getIsDisableSecond } from './utils'
 import type { ComputedRef, Ref, DefineComponent, UnwrapRef } from 'vue'
