+++ packages/varlet-ui/src/time-picker/clock.vue
@@ -39,7 +39,7 @@ import { computed, defineComponent, ref, watch } from 'vue'
 import dayjs from 'dayjs/esm'
 import { hoursAmpm, hours24, minSec } from './props'
 import { notConvert, convertHour, getIsDisableMinute, getIsDisableSecond, getNumberTime } from './utils'
-import { toNumber } from '../utils/shared'
+import { toNumber } from '@varlet/shared'
 import { createNamespace } from '../utils/components'
 import type { ComputedRef, Ref, PropType } from 'vue'
 import type { Time, AmPm, Format, AllowedTime } from './props'
