+++ packages/varlet-ui/src/uploader/Uploader.vue
@@ -77,7 +77,8 @@ import ImagePreview from '../image-preview'
 import Ripple from '../ripple'
 import { defineComponent, nextTick, reactive, computed, watch, ref } from 'vue'
 import { props } from './props'
-import { isNumber, isHTMLSupportImage, isHTMLSupportVideo, toNumber, isString } from '../utils/shared'
+import { isNumber, toNumber, isString } from '@varlet/shared'
+import { isHTMLSupportImage, isHTMLSupportVideo } from '../utils/shared'
 import { call, useValidation, createNamespace } from '../utils/components'
 import { useForm } from '../form/provide'
 import type { ComputedRef, Ref } from 'vue'
