+++ packages/varlet-ui/src/utils/shared.ts
@@ -1,5 +1,3 @@
-import { TimeData } from '../countdown/props'
-
 export interface CacheInstance<T> {
   cache: T[]
 
@@ -10,20 +8,6 @@ export interface CacheInstance<T> {
   remove(key: T): void
 
   clear(): void
-}
-
-export const toNumber = (val: number | string | boolean | undefined | null): number => {
-  if (val == null) return 0
-
-  if (isString(val)) {
-    val = parseFloat(val)
-    val = Number.isNaN(val) ? 0 : val
-    return val
-  }
-
-  if (isBool(val)) return Number(val)
-
-  return val
 }
 
 export const isHTMLSupportImage = (val: string | undefined | null) => {
@@ -42,62 +26,11 @@ export const isHTMLSupportVideo = (val: string | undefined | null) => {
   return val.startsWith('data:video') || /\.(mp4|webm|ogg)$/.test(val)
 }
 
-export const isString = (val: unknown): val is string => typeof val === 'string'
-
-export const isBool = (val: unknown): val is boolean => typeof val === 'boolean'
-
-export const isNumber = (val: unknown): val is number => typeof val === 'number'
-
-export const isPlainObject = (val: unknown): val is Record<string, any> =>
-  Object.prototype.toString.call(val) === '[object Object]'
-
-export const isObject = (val: unknown): val is Record<string, any> => typeof val === 'object' && val !== null
-
-export const isArray = (val: unknown): val is Array<any> => Array.isArray(val)
-
-export const isURL = (val: string | undefined | null) => {
-  if (!val) {
-    return false
-  }
-
-  return /^(http)|(\.*\/)/.test(val)
-}
-
-export const isEmpty = (val: unknown) =>
-  val === undefined || val === null || val === '' || (Array.isArray(val) && !val.length)
-
 export const removeItem = (arr: Array<unknown>, item: unknown) => {
   if (arr.length) {
     const index: number = arr.indexOf(item)
     if (index > -1) {
       return arr.splice(index, 1)
-    }
-  }
-}
-
-export const throttle = (method: any, mustRunDelay = 200): (() => void) => {
-  let timer: number
-  let start = 0
-
-  return function loop(this: unknown, ...args) {
-    const now = Date.now()
-    const elapsed = now - start
-
-    if (!start) {
-      start = now
-    }
-
-    if (timer) {
-      window.clearTimeout(timer)
-    }
-
-    if (elapsed >= mustRunDelay) {
-      method.apply(this, args)
-      start = now
-    } else {
-      timer = window.setTimeout(() => {
-        loop.apply(this, args)
-      }, mustRunDelay - elapsed)
     }
   }
 }
@@ -138,41 +71,4 @@ export const cubic = (value: number): number => value ** 3
 export const easeInOutCubic = (value: number): number =>
   value < 0.5 ? cubic(value * 2) / 2 : 1 - cubic((1 - value) * 2) / 2
 
-export function parseFormat(format: string, time: TimeData): string {
-  const scannedTimes = Object.values(time)
-  const scannedFormats = ['DD', 'HH', 'mm', 'ss']
-  const padValues = [24, 60, 60, 1000]
-
-  scannedFormats.forEach((scannedFormat, index) => {
-    if (!format.includes(scannedFormat)) {
-      scannedTimes[index + 1] += scannedTimes[index] * padValues[index]
-    } else {
-      format = format.replace(scannedFormat, String(scannedTimes[index]).padStart(2, '0'))
-    }
-  })
-
-  if (format.includes('S')) {
-    const ms = String(scannedTimes[scannedTimes.length - 1]).padStart(3, '0')
-
-    if (format.includes('SSS')) {
-      format = format.replace('SSS', ms)
-    } else if (format.includes('SS')) {
-      format = format.replace('SS', ms.slice(0, 2))
-    } else {
-      format = format.replace('S', ms.slice(0, 1))
-    }
-  }
-
-  return format
-}
-
 export const dt = (value: unknown, defaultText: string | undefined) => (value == null ? defaultText : value)
-
-export const inBrowser = () => typeof window !== 'undefined'
-
-export const uniq = (arr: Array<any>) => [...new Set(arr)]
-
-export function kebabCase(str: string): string {
-  const ret = str.replace(/([A-Z])/g, ' $1').trim()
-  return ret.split(' ').join('-').toLowerCase()
-}
